implementation of ACM Manager and EDM Manager as derived classes of common base
